requirements.txt
amqp==5.3.1
asgiref==3.8.1
async-timeout==5.0.1
billiard==4.2.1
celery==5.3.6
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.5
Django==4.2.7
django-anymail==13.0
django-celery-beat==2.8.1
django-cors-headers==4.3.1
django-timezone-field==7.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
idna==3.10
kombu==5.5.4
numpy==2.3.1
packaging==25.0
pandas==2.2.2
prompt_toolkit==3.0.51
psycopg2-binary==2.9.9
PyJWT==2.10.1
python-crontab==3.2.0
python-dateutil==2.9.0.post0
pytz==2025.2
redis==5.0.4
requests==2.32.4
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.13
